requirements.txt
biopython==1.76
numpy==1.18.2
PyYAML==5.3.1
